fix(payments-next): Move searchParams to params
