+++ apps/payments/next/app/[locale]/[offeringId]/checkout/layout.tsx
@@ -1,3 +1,7 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
 // TODO - Replace these placeholders as part of FXA-8227
 export const metadata = {
   title: 'Mozilla accounts',
@@ -5,10 +9,8 @@ export const metadata = {
 };
 
 export interface CheckoutSearchParams {
-  interval?: string;
-  promotion_code?: string;
   experiment?: string;
-  locale?: string;
+  promotion_code?: string;
 }
 
 export default function RootLayout({
