+++ apps/payments/next/app/[locale]/[offeringId]/checkout/new/page.tsx
@@ -2,9 +2,6 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-/* eslint-disable @typescript-eslint/no-explicit-any */
-declare module '*.svg' {
-  const content: any;
-  export const ReactComponent: any;
-  export default content;
+export default async function CheckoutNew() {
+  return <></>;
 }
