--- apps/payments/next/app/[offeringId]/checkout/new/page.tsx
@@ -1,3 +1,0 @@
-export default async function CheckoutNew() {
-  return <></>;
-}
