+++ apps/payments/next/app/api/auth/[...nextauth]/route.ts
@@ -1,1 +1,5 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
 export { GET, POST } from '../../../../auth';
