+++ apps/payments/next/app/layout.tsx
@@ -1,3 +1,7 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
 import Image from 'next/image';
 // TODO - Move shared SVGs to a shared library in fxa/libs/shared/*
 import mozillaLogo from '../images/moz-logo-bw-rgb.svg';
