+++ apps/payments/next/app/page.tsx
@@ -1,3 +1,7 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
 import Link from 'next/link';
 
 export default function Index() {
@@ -14,7 +18,7 @@ export default function Index() {
           <h2>123Done - Monthly</h2>
           <Link
             className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded"
-            href="/123done/checkout/b6115e72-2a3f-4de8-a58a-e231bfeea85d?interval=monthly"
+            href="/en/123done/checkout/monthly/b6115e72-2a3f-4de8-a58a-e231bfeea85d"
           >
             Redirect
           </Link>
@@ -23,7 +27,7 @@ export default function Index() {
           <h2>123Done - Yearly</h2>
           <Link
             className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded"
-            href="/123done/checkout/b6115e72-2a3f-4de8-a58a-e231bfeea85d?interval=yearly"
+            href="/en/123done/checkout/yearly/b6115e72-2a3f-4de8-a58a-e231bfeea85d"
           >
             Redirect
           </Link>
