+++ libs/payments/ui/src/index.ts
@@ -1,1 +1,7 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
 // Use this file to export React client components (e.g. those with 'use client' directive) or other non-server utilities
+
+export * from './lib/utils/helpers';
