+++ libs/payments/ui/src/server.ts
@@ -1,3 +1,7 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
 // Use this file to export React server components
 export * from './lib/server/purchase-details';
 export * from './lib/server/terms-and-privacy';
